Don't store empty segments.  When inserting empty strings, the code was writing out a segment made up of a single leaf node containing the \0 header.  LeafReader assumed that leaf nodes always contained at least one term, so assertions would fail.
While it would be possible to support reading and merging empty
segments, there's no reason to do so.  While this change could have
been done in writeZeroSegment(), I put it in leafWriterFlush() so that
it would work right if segmentMerge() created an empty segment, which
could happen with future changes to how deleted documents are handled. (CVS 3484)